#23 refactor: Reorganização da estrutura do projeto
Move arquivos de componentes, serviços, configuração e estilos para diretórios mais apropriados com base em funcionalidade e reusabilidade.
