+++ package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@tanstack/react-query": "^5.76.2",
     "axios": "^1.9.0",
     "react": "^18.3.1",
     "react-beautiful-dnd": "^13.1.1",
@@ -17,6 +18,7 @@
     "react-icons": "^5.5.0",
     "react-router-dom": "^7.5.3",
     "react-toastify": "^11.0.5",
+    "react-window": "^1.8.11",
     "taskflow-frontend": "file:"
   },
   "devDependencies": {
