+++ src/App.jsx
@@ -1,29 +1,44 @@
 import { Routes, Route } from 'react-router-dom';
-import Login from './pages/login';
-import ProtectedRoute from './components/ProtectedRoute';
-import BoardWorkspace from './pages/boards/BoardWorkspace';
-import './App.css';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { Suspense, lazy } from 'react';
+import './styles/App.css';
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
+const Login = lazy(() => import('./pages/login'));
+const BoardWorkspace = lazy(() => import('./pages/boards/BoardWorkspace'));
+const ProtectedRoute = lazy(() => import('./components/ProtectedRoute'));
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      staleTime: 1000 * 60 * 5, // 5 minutes
+      cacheTime: 1000 * 60 * 30, // 30 minutes
+      retry: 1,
+      refetchOnWindowFocus: false,
+    },
+  },
+});
 
 function App() {
   return (
-    <>
-    <Routes>
-      <Route path="/" element={<Login />} />
-      <Route path="/login" element={<Login />} />
-      <Route
-        path="/boards"
-        element={
-          <ProtectedRoute>
-            <BoardWorkspace />
-          </ProtectedRoute>
-        }
-      />
-    </Routes>
-    <ToastContainer position="top-right" autoClose={3000} />
-    </>
+    <QueryClientProvider client={queryClient}>
+      <Suspense fallback={<div>Carregando...</div>}>
+        <Routes>
+          <Route path="/" element={<Login />} />
+          <Route path="/login" element={<Login />} />
+          <Route
+            path="/boards"
+            element={
+              <ProtectedRoute>
+                <BoardWorkspace />
+              </ProtectedRoute>
+            }
+          />
+        </Routes>
+      </Suspense>
+      <ToastContainer position="top-right" autoClose={3000} />
+    </QueryClientProvider>
   );
 }
