+++ src/config/config.js
@@ -1,2 +1,1 @@
-// src/config.js
 export const VITE_API_URL = import.meta.env.VITE_API_URL;
