+++ src/pages/boards/components/modals/DeleteConfirmModal.jsx
