+++ src/pages/login/LoginForm.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import styles from "./styles/styles.module.css";
-import { login } from "../../services/authService";
+import { login } from "../../api/authService";
 import RegisterModal from "./RegisterModal"; // Importa o modal de registro
-import { register } from "../../services/registerService";
+import { register } from "../../api/registerService";
 
 
 const LoginForm = () => {
